libemu  - emt64 fixes to allow compiling on x86_64 plattforms    maybe it is a bad habbit to store integer values on pointers to save some bytes    in this case I stored eip in a hashmap, using the hashmaps key pointer of type void * as store    compiling the code on a 64bit plattform .. made the cast invalid, as sizeof(void *) != sizeof(uint32_t)    therefore I had to cast to uintptr_t before casting to uint32_t in some places.    another thing is printing memory addresses, %08x works fine on 32bit plattforms, to be portable one should use %p, which does not prepend leading zeros.

+++ src/emu_hashtable.c
@@ -257,14 +257,14 @@ bool emu_hashtable_string_cmp(void *a, void *b)
 
 uint32_t emu_hashtable_ptr_hash(void *key)
 {
-	uint32_t ukey = (uint32_t)key;
+	uint32_t ukey = (uint32_t)(uintptr_t)key;
 	ukey++;
 	return ukey;
 }
 
 bool emu_hashtable_ptr_cmp(void *a, void *b)
 {
-	if ( (uint32_t)a == (uint32_t)b )
+	if ( (uintptr_t)a == (uintptr_t)b )
 		return true;
 
 	return false;
